requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
class-registry==2.1.2
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filters==1.3.2
filters-django==1.0.5
gunicorn==20.1.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
six==1.16.0
sqlparse==0.4.2
python-dotenv~=0.19.2
# django-heroku==0.3.0

